Testing /pg_activity_stats json
This is a pretty tricky test because it is hard to insert records into a
test db to test what this json will be. So I ended up stubbing the whole
active record part.

It is also odd that OpenStruct.to_json puts all attributes under
'table', so I needed to patch that for this test also.

...Overall I'm not a huge fan of this test...