Move from /spec to /test
I am using Test::Unit::TestCase and that seemed like it should be in a
test dir not spec, since it is a sub-class of 'TestCase'. There is a
little bit of confusion about this since the Sinatra page calls these
specs but mini test seems to call them tests.